Update default.php
I found a problem with the CSS code: the font color of the title is not taken into account because the CSS code applied to the first line of the table (line 30 to 34) is bypassed by the code block that follows (which manages the background color alternation of the rows in the table).

There are 2 alternatives: move the CSS code (as in the original module) or then consider the use of thead tags, th and tbody.

In addition, I would like to suggest using javascript that makes it very easy to sort the tables. We could even add this feature with an option (we should add the option in the XML file: see where you think it is the most relevant. Of course, the js file should be added to the ad hoc file. You will see that I started to code this by commenting on the condition to load the javascript file.

Here are these proposals:
- modified CSS code to embed the tag th,
- add condition to integrate thead, th and tbody,
- add script to sort the table (thead is mandatory to use this script)