Update default.php : problème en cas d'ajout d'un suffixe de class dans le module
I had forgotten about this problem in the module: in case of adding one or more classes, there is no . in front of the classes and these are not correctly written in the style definition. This results in an error and affected styles are not applied.

I patched the mostafa's module by replacing '$moduleclass_sfx' with 'str_replace(' ','.',$moduleclass_sfx)' except on line 88.
My patch has one problem: the first (or only) class must start with a space.